copy changes from keepitfresh repo

+++ src/main/webapp/WEB-INF/views/list-settings.jsp
@@ -1,7 +1,7 @@
 <%@ include file="common/header.jspf"%>
 <%@ include file="common/navigation.jspf"%>
 
-<div class="container">
+<div class="container" id="settings">
     <table class="table table-striped">
         <thead>
             <tr>
@@ -13,10 +13,10 @@
         <tbody>
             <c:forEach items="${settings}" var="setting">
                 <tr>
-                    <td>${setting.emailAddress}</td>
-                    <td>${setting.daysBeforeExp}</td>
-                    <td><a type="button" class="btn btn-primary" href="/update-setting?id=${setting.id}">Edit</a> 
-                        <a type="button" class="btn btn-danger" href="/delete-setting?id=${setting.id}">Delete</a>
+                    <td class="col-md-3">${setting.emailAddress}</td>
+                    <td class="col-md-4">${setting.daysBeforeExp}</td>
+                    <td class="col-md-5"><a type="button" class="btn btn-primary" href="/update-setting?id=${setting.id}">Edit</a> 
+                        <a type="button" class="btn btn-danger" href="/delete-setting?id=${setting.id}" data-toggle="confirmation">Delete</a>
                     </td>
                 </tr>
             </c:forEach>
@@ -38,8 +38,9 @@
         <tbody>
             <c:forEach items="${categories}" var="category">
                 <tr>
-                    <td>${category.name}</td>
-                    <td><a type="button" class="btn btn-primary" href="/update-category?id=${category.id}">Edit</a> 
+                    <td class="col-md-7">${category.name}</td>
+                    <td class="col-md-5">
+                    	<a type="button" class="btn btn-primary" href="/update-category?id=${category.id}">Edit</a> 
                         <a type="button" class="btn btn-danger" href="/delete-category?id=${category.id}">Delete</a>
                     </td>
                 </tr>
